Merge 1.12 staging

+++ app/utils/file.js
@@ -1,8 +1,8 @@
-// Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
-// See License.txt for license information.
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
 
 import {Platform} from 'react-native';
-import RNFetchBlob from 'react-native-fetch-blob';
+import RNFetchBlob from 'rn-fetch-blob';
 import mimeDB from 'mime-db';
 
 import {lookupMimeType} from 'mattermost-redux/utils/file_utils';
@@ -88,9 +88,20 @@ export async function getFileCacheSize() {
 }
 
 export async function deleteFileCache() {
-    await RNFetchBlob.fs.unlink(DOCUMENTS_PATH);
-    await RNFetchBlob.fs.unlink(IMAGES_PATH);
-    await RNFetchBlob.fs.unlink(VIDEOS_PATH);
+    const isDocsDir = await RNFetchBlob.fs.isDir(DOCUMENTS_PATH);
+    const isImagesDir = await RNFetchBlob.fs.isDir(IMAGES_PATH);
+    const isVideosDir = await RNFetchBlob.fs.isDir(VIDEOS_PATH);
+    if (isDocsDir) {
+        await RNFetchBlob.fs.unlink(DOCUMENTS_PATH);
+    }
+
+    if (isImagesDir) {
+        await RNFetchBlob.fs.unlink(IMAGES_PATH);
+    }
+
+    if (isVideosDir) {
+        await RNFetchBlob.fs.unlink(VIDEOS_PATH);
+    }
     return true;
 }
 
@@ -214,4 +225,12 @@ function populateMaps() {
             types[extension] = type;
         }
     });
+}
+
+export function getLocalFilePathFromFile(dir, file) {
+    if (dir && file && file.caption && file.data && file.data.id) {
+        return `${dir}/${file.data.id}-${decodeURIComponent(file.caption).replace(/\s+/g, '-')}`;
+    }
+
+    return null;
 }
